[iOS] chore: 파일 정리 Main 그룹 삭제, MainView, TokenCellView, TokenEditView로 분리 MainUsecase.swift, TokenDropDelegate.swift, NavigationFlowObject.swift 추가 common에 Extension, Util 포함 유틸 추가 NavigationFlowObject.swift,  NavigationLazyView 추가
