+++ iOS/DaDaIkSeon/DaDaIkSeon/Common/Util/NavigationFlowObject.swift
@@ -1,0 +1,12 @@
+//
+//  NavigationFlowObject.swift
+//  DaDaIkSeon
+//
+//  Created by 정재명 on 2020/12/06.
+//
+
+import Foundation
+
+class NavigationFlowObject: ObservableObject {
+    @Published var isActive = false
+}
