+++ iOS/DaDaIkSeon/DaDaIkSeon/MainView/View/MainView.swift
@@ -8,40 +8,6 @@
 import SwiftUI
 import UniformTypeIdentifiers
 
-struct MainState {
-    var service: TokenServiceable
-    var filteredTokens: [Token]
-    var isSearching: Bool
-    var mainToken: Token
-    var checkBoxMode: Bool
-    var selectedTokens: [UUID: Bool]
-    var settingMode: Bool
-    var selectedCount: Int
-    var zeroTokenState: Bool
-    var tokenOnDrag: Token?
-}
-
-enum MainInput {
-    case search(_ text: String)
-    case startSearch
-    case endSearch
-    case showCheckBox
-    case hideCheckBox
-    case selectCell(_ id: UUID)
-    case startSetting
-    case endSetting
-    case deleteSelectedTokens
-    case refreshTokens
-    case moveToMain(_ id: UUID)
-    case move(_ from: Int, _ target: Int)
-    case startDragging(_ token: Token)
-    case endDragging
-}
-
-class NavigationFlowObject: ObservableObject {
-    @Published var isActive = false
-}
-
 struct MainView: View {
     
     // MARK: ViewModel
@@ -202,54 +168,6 @@ struct MainView_Previews: PreviewProvider {
     static var previews: some View {
         let service = TokenService(StorageManager())
         MainView(service: service)
-    }
-    
-}
-
-struct TokenDropDelegate: DropDelegate {
-    
-    private let item: Token
-    @Binding private var tokenOnDrag: Token?
-    private var service: TokenServiceable
-    private var moveAction: (Int, Int) -> Void
-    private var endAction: () -> Void
-    private var listData: [Token] { service.tokenList() }
-    
-    init(item: Token,
-         tokenOnDrag: Binding<Token?>,
-         service: TokenServiceable,
-         moveAction: @escaping (Int, Int) -> Void,
-         endAction: @escaping () -> Void) {
-        self.item = item
-        self._tokenOnDrag = tokenOnDrag
-        self.service = service
-        self.moveAction = moveAction
-        self.endAction = endAction
-    }
-    
-    func dropEntered(info: DropInfo) {
-        
-        // item은 원래 위치, tokenOnDrag는 현재 드래그 중인 토큰의 위치
-        if item != tokenOnDrag {
-            guard let from = listData.firstIndex(where: {
-                $0 == tokenOnDrag
-            }) else { return }
-            guard let target = listData.firstIndex(where: {
-                $0 == item
-            }) else { return }
-            if listData[target].id != listData[from].id {
-                moveAction(from, target)
-            }
-        }
-    }
-    
-    func dropUpdated(info: DropInfo) -> DropProposal? {
-        return DropProposal(operation: .move)
-    }
-    
-    func performDrop(info: DropInfo) -> Bool {
-        endAction()
-        return true
     }
     
 }
